Use page template structure for location pages.
In order to display previews in the CMS for custom page templates, we
need to use the same page and export template structure as the other
pages. This also renames location to location page and updates references
accordingly.
